Prepare campaign for C22_WMDE_Mobile_Test_08
