+++ banners/mobile/components/Banner.jsx
@@ -13,10 +13,16 @@ import FundsDistributionAccordion from '../../../components/UseOfFunds/FundsDist
 import { BannerType } from '../../../shared/BannerType';
 import SlideState from '../../../shared/slide_state';
 import createDynamicCampaignText from '../../../shared/create_dynamic_campaign_text';
+import SoftClose from '../../../components/SoftClose/SoftClose';
+import CssTransition from '../../../shared/css_transition';
 
-const PENDING = 0;
-const VISIBLE = 1;
-const CLOSED = 2;
+const BannerVisibilityState = Object.freeze( {
+	PENDING: Symbol( 'pending' ),
+	MINI_VISIBLE: Symbol( 'mini-visible' ),
+	FULL_VISIBLE: Symbol( 'full-visible' ),
+	SOFT_CLOSING: Symbol( 'soft-closing' ),
+	CLOSED: Symbol( 'closed' )
+} );
 
 const SLIDESHOW_START_DELAY = 2000;
 
@@ -37,8 +43,7 @@ export default class Banner extends Component {
 	constructor( props ) {
 		super( props );
 		this.state = {
-			displayState: PENDING,
-			isFullPageVisible: false,
+			bannerVisibilityState: BannerVisibilityState.PENDING,
 			isFundsModalVisible: false,
 			textHighlight: HighlightState.WAITING
 		};
@@ -62,11 +67,13 @@ export default class Banner extends Component {
 
 	miniBannerTransitionRef = createRef();
 	fullBannerTransitionRef = createRef();
+	softCloseTransitionRef = createRef();
+	softCloseRef = createRef();
 
 	componentDidMount() {
 		this.props.registerDisplayBanner(
 			() => {
-				this.setState( { displayState: VISIBLE } );
+				this.setState( { bannerVisibilityState: BannerVisibilityState.MINI_VISIBLE } );
 				this.slideInBanner();
 			}
 		);
@@ -84,16 +91,18 @@ export default class Banner extends Component {
 	}
 
 	setContentSize() {
-		if ( this.state.displayState !== VISIBLE ) {
-			return;
-		}
-
-		if ( this.state.isFullPageVisible ) {
-			this.props.skinAdjuster.addSpaceInstantly( this.getFullBannerHeight() );
-			this.fullPageBannerReRender();
-		} else {
-			this.props.skinAdjuster.addSpaceInstantly( this.getMiniBannerHeight() );
-			this.adjustFollowupBannerHeight( this.getMiniBannerHeight() );
+		switch ( this.state.bannerVisibilityState ) {
+			case BannerVisibilityState.MINI_VISIBLE:
+				this.props.skinAdjuster.addSpaceInstantly( this.getMiniBannerHeight() );
+				this.adjustFollowupBannerHeight( this.getMiniBannerHeight() );
+				break;
+			case BannerVisibilityState.FULL_VISIBLE:
+				this.props.skinAdjuster.addSpaceInstantly( this.getFullBannerHeight() );
+				this.fullPageBannerReRender();
+				break;
+			case BannerVisibilityState.SOFT_CLOSING:
+				this.props.skinAdjuster.addSpaceInstantly( this.getSoftCloseBannerHeight() );
+				break;
 		}
 	}
 
@@ -102,12 +111,11 @@ export default class Banner extends Component {
 		this.showFullPageBanner( e );
 	};
 
-	// eslint-disable-next-line no-unused-vars
-	showFullPageBanner = e => {
+	showFullPageBanner = () => {
 		this.stopSliderAutoplay();
 		window.scrollTo( 0, 0 );
 		this.transitionToFullpage( this.getMiniBannerHeight() );
-		this.setState( { isFullPageVisible: true } );
+		this.setState( { bannerVisibilityState: BannerVisibilityState.FULL_VISIBLE } );
 	};
 
 	toggleFundsModal = e => {
@@ -143,6 +151,10 @@ export default class Banner extends Component {
 		return this.fullBannerTransitionRef.current ? this.fullBannerTransitionRef.current.base.offsetHeight : 0;
 	}
 
+	getSoftCloseBannerHeight() {
+		return this.softCloseTransitionRef.current ? this.softCloseTransitionRef.current.offsetHeight : 0;
+	}
+
 	onSlideChange = ( index ) => {
 		this.slideState.onSlideChange( index );
 	};
@@ -152,13 +164,36 @@ export default class Banner extends Component {
 		this.stopSliderAutoplay = onStopAutoplay;
 	};
 
-	closeBanner = e => {
+	onSoftCloseMiniBanner = e => {
 		e.preventDefault();
-		this.setState( {
-			displayState: CLOSED,
-			isFullPageVisible: false
-		} );
+		this.softCloseRef.current?.startProgress();
+		this.setState(
+			{ bannerVisibilityState: BannerVisibilityState.SOFT_CLOSING },
+			() => this.setContentSize()
+		);
+	};
+
+	onCloseFullBanner = e => {
+		e.preventDefault();
+		this.setState( { bannerVisibilityState: BannerVisibilityState.CLOSED } );
+		this.props.onClose( 'banner-closed-full' );
+	};
+
+	onMaybeLater = e => {
+		e.preventDefault();
+		this.setState( { bannerVisibilityState: BannerVisibilityState.CLOSED } );
+		this.props.onMaybeLater();
+	};
+
+	onCloseBanner = e => {
+		e.preventDefault();
+		this.setState( { bannerVisibilityState: BannerVisibilityState.CLOSED } );
 		this.props.onClose();
+	};
+
+	onTimeOutClose = () => {
+		this.setState( { bannerVisibilityState: BannerVisibilityState.CLOSED } );
+		this.props.onClose( 'micro-banner-ignored', new CssTransition( 1000 ) );
 	};
 
 	registerBannerTransition = cb => { this.slideInBanner = cb; };
@@ -200,10 +235,11 @@ export default class Banner extends Component {
 		const campaignProjection = props.campaignProjection;
 		return <div className={classNames( {
 			'wmde-banner': true,
-			'wmde-banner--hidden': state.displayState === CLOSED,
-			'wmde-banner--visible': state.displayState === VISIBLE,
-			'wmde-banner--mini-banner': !state.isFullPageVisible,
-			'wmde-banner--full-page': state.isFullPageVisible,
+			'wmde-banner--hidden': state.bannerVisibilityState === BannerVisibilityState.CLOSED,
+			'wmde-banner--visible': state.bannerVisibilityState !== BannerVisibilityState.CLOSED && state.bannerVisibilityState !== BannerVisibilityState.PENDING,
+			'wmde-banner--soft-closing': state.bannerVisibilityState === BannerVisibilityState.SOFT_CLOSING,
+			'wmde-banner--mini-banner': state.bannerVisibilityState === BannerVisibilityState.MINI_VISIBLE,
+			'wmde-banner--full-page': state.bannerVisibilityState === BannerVisibilityState.FULL_VISIBLE,
 			'wmde-banner--animate-highlight': state.textHighlight === HighlightState.ANIMATE,
 			'wmde-banner--ctrl': props.bannerType === BannerType.CTRL,
 			'wmde-banner--var': props.bannerType === BannerType.VAR
@@ -219,7 +255,7 @@ export default class Banner extends Component {
 				>
 					<MiniBanner
 						{ ...props }
-						onClose={ this.closeBanner }
+						onClose={ this.onSoftCloseMiniBanner }
 						campaignProjection={ campaignProjection }
 						setStartAnimation={ this.registerStartProgressBarInMiniBanner }
 						onExpandFullpage={ this.showFullPageBannerFromMiniBanner }
@@ -245,7 +281,7 @@ export default class Banner extends Component {
 						onSubmitRecurring={ () => props.onSubmit( 'submit-recurring' ) }
 						onSubmitNonRecurring={ () => props.onSubmit( 'submit-non-recurring' ) }
 						onChangeToYearly={ this.onDonationFormChangeToYearly }
-						onClose={ this.closeBanner }
+						onClose={ this.onCloseFullBanner }
 						campaignProjection={ campaignProjection }
 						donationForm={props.donationForm}
 						setStartAnimation={ this.registerStartProgressBarInFullPageBanner }
@@ -253,6 +289,18 @@ export default class Banner extends Component {
 						dynamicCampaignText={ this.dynamicCampaignText }
 					/>
 				</FollowupTransition>
+				<div
+					ref={ this.softCloseTransitionRef }
+					className="soft-close-container banner-position--state-finished banner-position banner-position--fixed"
+					style="top: 0px"
+				>
+					<SoftClose
+						onMaybeLater={ this.onMaybeLater }
+						onCloseBanner={ this.onCloseBanner }
+						onTimeOutClose={ this.onTimeOutClose }
+						ref={ this.softCloseRef }
+					/>
+				</div>
 			</TranslationContext.Provider>
 			<FundsModal
 				isFundsModalVisible={ this.state.isFundsModalVisible }
