+++ banners/mobile/styles/styles.pcss
@@ -29,6 +29,7 @@ weight should be component < theme < banner.
 @import '../../../components/DonationForm/2-step-form.pcss';
 @import '../../../components/Slider/Slider.pcss';
 @import '../../../components/ProgressBar/ProgressBar.pcss';
+@import '../../../components/SoftClose/SoftClose.pcss';
 @import '../../../components/UseOfFunds/UseOfFunds.pcss';
 @import '../../../components/UseOfFunds/FundsDistributionAccordion.pcss';
 
@@ -43,3 +44,4 @@ weight should be component < theme < banner.
 @import '../../../themes/gohan/DonationForm/2-step-form.pcss';
 @import '../../../themes/gohan/Slider/Slider.pcss';
 @import '../../../themes/gohan/ProgressBar/ProgressBar.pcss';
+@import '../../../themes/gohan/SoftClose/SoftClose.pcss';
