--- banners/mobile/styles/styles_var.pcss
@@ -1,47 +1,0 @@
-/**
-The order in which imports happen is important because component ones
-can be overridden by the theme and the individual banners. The variable
-weight should be component < theme < banner.
- */
-
-/* COMPONENT VARIABLES */
-@import '../../../components/ButtonClose/ButtonCloseVariables.pcss';
-@import '../../../components/Footer/FooterVariables.pcss';
-@import '../../../components/UseOfFunds/UseOfFundsVariables.pcss';
-
-@import '../../../themes/gohan/theme-variables.pcss';
-@import './variables.pcss';
-
-@import './Banner.pcss';
-@import './MiniBanner.pcss';
-@import './FullBanner.pcss';
-
-/* COMPONENTS */
-@import '../../../components/BannerTransition/BannerTransition.pcss';
-@import '../../../components/BannerTransition/FollowupTransition.pcss';
-@import '../../../components/Footer/Footer.pcss';
-@import '../../../components/Footer/SelectionInput.pcss';
-@import '../../../components/ButtonClose/ButtonClose.pcss';
-@import '../../../components/DonationForm/DonationForm.pcss';
-@import '../../../components/DonationForm/SelectGroup.pcss';
-@import '../../../components/DonationForm/SelectCustomAmount.pcss';
-@import '../../../components/DonationForm/SmsBox.pcss';
-@import '../../../components/DonationForm/2-step-form.pcss';
-@import '../../../components/Slider/Slider.pcss';
-@import '../../../components/ProgressBar/ProgressBar.pcss';
-@import '../../../components/SoftClose/SoftClose.pcss';
-@import '../../../components/UseOfFunds/UseOfFunds.pcss';
-@import '../../../components/UseOfFunds/FundsDistributionAccordion.pcss';
-
-/* THEME */
-@import '../../../themes/gohan/animated-text-highlight.pcss';
-@import '../../../themes/gohan/DonationForm/DonationForm.pcss';
-@import '../../../themes/gohan/Footer/Footer.pcss';
-@import '../../../themes/gohan/Footer/SelectionInput.pcss';
-@import '../../../themes/gohan/DonationForm/SelectGroup.pcss';
-@import '../../../themes/gohan/DonationForm/SelectCustomAmount.pcss';
-@import '../../../themes/gohan/DonationForm/SmsBox.pcss';
-@import '../../../themes/gohan/DonationForm/2-step-form.pcss';
-@import '../../../themes/gohan/Slider/Slider.pcss';
-@import '../../../themes/gohan/ProgressBar/ProgressBar.pcss';
-@import '../../../themes/gohan/SoftClose/SoftClose.pcss';
